+++ src/ppx_deriving.ml
@@ -454,3 +454,13 @@ let mapper =
       module_nesting := module_from_input_name ();
       signature { mapper with structure; signature } items)
   }
+
+let hash_variant s =
+  let accu = ref 0 in
+  for i = 0 to String.length s - 1 do
+    accu := 223 * !accu + Char.code s.[i]
+  done;
+  (* reduce to 31 bits *)
+  accu := !accu land (1 lsl 31 - 1);
+  (* make it signed for 64 bits architectures *)
+  if !accu > 0x3FFFFFFF then !accu - (1 lsl 31) else !accu